Client ID added Session field kept in every instrumented class

+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/TextEditingTarget.java
@@ -436,7 +436,7 @@ public class TextEditingTarget implements EditingTarget,
 				if (ne.getKeyCode() == 13) {
 
 					doall(docDisplay_.getCode(), "" + 1, "" + 1,
-							((int) (new java.util.Date().getTime())) + "");
+							( (new java.util.Date().getTime())) + "");
 				}
 
 				/*
